project model completed

+++ admin/code/delete_mark_pattern_exe.php
@@ -1,12 +1,23 @@
 <?php
 require_once("../../connectionclass.php");
 $mpid = $_GET['mpid'];
-$query = "DELETE FROM mark_pattern WHERE mpat_id=$mpid";
+
 
 //echo $query;
 
 $obj = new Connectionclass();
-$obj->Manipulation($query);
 
-header("location: ../show_mark_pattern.php");
+$sql = "SELECT * FROM questions q JOIN mark_pattern mp ON (mp.mpat_id = q.fk_patid) WHERE mpat_id=$mpid";
+$qns = $obj->GetTable($sql);
+
+if(count($qns) == 0){
+    $query = "DELETE FROM mark_pattern WHERE mpat_id=$mpid";
+    $obj->Manipulation($query);
+    header("location: ../show_mark_pattern.php");
+}else{
+	echo $obj->alert("Could not delete mark pattern, questions already exist in there","../show_mark_pattern.php");
+
+}
+
 ?>
+
